int to float change in GC

+++ hw3/fastq_filtrator_funcs.py
@@ -94,7 +94,7 @@ if __name__ == "__main__":
     )
     input_path = input("Enter path to the input file: ")
     output_path = input("Enter the prefix of the output file: ")
-    gc_bounds_in = tuple(int(i) for i in input("Enter the bounds for GC: ").split())
+    gc_bounds_in = tuple(float(i) for i in input("Enter the bounds for GC: ").split())
     length_bounds_in = tuple(int(i) for i in input("Enter the bounds for length of the read: ").split())
     qs_threshold_in = input("Enter the minimal mean quality of the read: ")
     try:
